+++ config/dev.exs
@@ -1,3 +1,0 @@
-import Config
-
-config :redis_mutex, redis_url: "redis://localhost:6379"