+++ test/test_helper.exs
@@ -1,8 +1,2 @@
-ExUnit.configure(exclude: [:redis_dependent])
 ExUnit.start()
-
-{:ok, files} = File.ls("./test/support")
-
-Enum.each(files, fn file ->
-  Code.require_file("support/#{file}", __DIR__)
-end)
+ExUnit.configure(exclude: [:skip])