+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## 1.0.0 (2024-02-12)
+
+### Changed
+
+**Breaking changes**
+- `RedisMutex` no longer starts as its own application. Instead, it can re-use an existing Redis connection
+  or be started in your application's supervision tree. Here is an example of starting it in an application's
+  supervision tree:
+```elixir
+  @impl Application
+  def start(_type, _args) do
+    children = other_children() ++ [{RedisMutex, redis_url: System.get_env("REDIS_URL")}]
+    Supervisor.start_link(children, strategy: :one_for_one, name: MyApp.Supervisor)
+  end
+```
+  Please see the README for more details.
+- `use RedisMutex` replaced in favor of calling `RedisMutex.with_lock/3` directly
+- `with_lock` changed to take a function argument instead of a do block to perform
+- `with_lock` changed to take a keyword list of options instead of optional `timeout` and `expiry` arguments
+
+
 ## 0.6.0 (2023-11-08)
 
 ### Changed